+++ source/utils.cpp
@@ -1,4 +1,5 @@
 #include "utils.h"
+#include <stdint.h>
 
 #ifndef __3DS__
 #define MAX_TOP_WIDTH __UINT64_MAX__
@@ -17,9 +18,9 @@ std::string strip(std::string s) {
 std::string char_vec_to_string(std::vector<char> &line) {
 
   std::string temp_str = "";
-  int letters = 0;
+  uint64_t letters = 0;
   for (const auto &ch : line) {
-    if ((uint64_t)letters != MAX_TOP_WIDTH) {
+    if (letters != MAX_TOP_WIDTH) {
       // Store characters to display
       temp_str.push_back(ch);
       letters++;
